code/requirements.txt
azure-ai-inference>=1.0.0b9
azure-core>=1.30.0
azure-search-documents>=11.4.0
azure-identity>=1.15.0
anthropic>=0.18.1
chardet>=5.2.0
openai>=1.12.0
google-genai>=0.7.1
google-cloud-aiplatform>=1.38.0
jsonschema>=4.19.1
mcp>=1.0.0
marshmallow<4.0.0
packaging<24,>=16.8
protobuf<5,>=3.20
matplotlib>=3.8.0
python-dotenv>=1.0.0
pymilvus>=1.1.0
qdrant-client>=1.14.0
aiohttp>=3.9.1
pyyaml>=6.0.1
feedparser>=6.0.1
httpx>=0.28.1
huggingface_hub>=0.31.0
seaborn>=0.13.0
python-dotenv>=1.0.0
